% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Calculate mean of sparse numeric vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments}
}
\value{
Numeric value representing the mean
}
\description{
Calculate mean of sparse numeric vector
}
