% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{mean}
\alias{mean}
\alias{mean,sparse_numeric-method}
\title{Mean of a sparse vector}
\usage{
mean(x, ...)

\S4method{mean}{sparse_numeric}(x)
}
\arguments{
\item{x}{The sparse vector}

\item{...}{Not Necessary}
}
\value{
numeric number
}
\description{
This function takes the mean of a sparse vector, accounting for the 0s not listed in the sparse vector
}
