% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{norm,sparse_numeric-method}
\alias{norm,sparse_numeric-method}
\title{Calculate squared norm of sparse numeric vector}
\usage{
\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments}
}
\value{
Numeric value representing the L2 norm
}
\description{
Calculate squared norm of sparse numeric vector
}
