% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{norm}
\alias{norm}
\title{Calculate norm of a vector}
\usage{
norm(x, ...)
}
\arguments{
\item{x}{An object for which the norm should be calculated}

\item{...}{Additional arguments passed to methods}
}
\value{
Numeric value representing the norm
}
\description{
Calculate norm of a vector
}
