% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{+,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\title{Addition operator for sparse_numeric vectors}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{A sparse_numeric object}

\item{e2}{A sparse_numeric object}
}
\value{
A sparse_numeric object
}
\description{
Addition operator for sparse_numeric vectors
}
