% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{sort,sparse_numeric-method}
\alias{sort,sparse_numeric-method}
\title{Sort sparse_numeric vector}
\usage{
\S4method{sort}{sparse_numeric}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{decreasing}{values decrease}

\item{...}{Additional arguments}
}
\value{
A sparse_numeric object with sorted values
}
\description{
Sort sparse_numeric vector
}
