% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Add two sparse numeric vectors}
\usage{
sparse_add(x, y, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional arguments (not used)}
}
\value{
A sparse_numeric object containing the sum
}
\description{
Add two sparse numeric vectors
}
\section{Functions}{
\itemize{
\item \code{sparse_add(x = sparse_numeric, y = sparse_numeric)}: Method for sparse_numeric vectors

}}
\examples{
x <- as(c(1, 0, 3, 0, 5), "sparse_numeric")
y <- as(c(0, 2, 0, 4, 0), "sparse_numeric")
result <- sparse_add(x, y)
}
