% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Cross product of two sparse numeric vectors}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional arguments (not used)}
}
\value{
A numeric scalar representing the cross product
}
\description{
Cross product of two sparse numeric vectors
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_crossprod(x = sparse_numeric, y = sparse_numeric)}: Method for sparse_numeric vectors

}}
\examples{
x <- as(c(1, 0, 3, 0, 5), "sparse_numeric")
y <- as(c(2, 0, 4, 0, 6), "sparse_numeric")
result <- sparse_crossprod(x, y)
}
