% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{sparse_div}
\alias{sparse_div}
\alias{sparse_div,sparse_numeric,sparse_numeric-method}
\title{divides a sparse vector by another sparse vector}
\usage{
sparse_div(first, second, ...)

\S4method{sparse_div}{sparse_numeric,sparse_numeric}(first, second)
}
\arguments{
\item{first}{The first sparse vector}

\item{second}{The second sparse vector, of the same length as the first}

\item{...}{Not Necessary}
}
\value{
Sparse vector
}
\description{
This function returns the product of dividing one sparse vector by another sparse vector of the same length
}
