% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiply two sparse numeric vectors}
\usage{
sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional arguments (not used)}
}
\value{
A sparse_numeric object containing the product
}
\description{
Multiply two sparse numeric vectors
}
\section{Functions}{
\itemize{
\item \code{sparse_mult(x = sparse_numeric, y = sparse_numeric)}: Method for sparse_numeric vectors

}}
\examples{
x <- as(c(1, 0, 3, 0, 5), "sparse_numeric")
y <- as(c(2, 2, 2, 2, 2), "sparse_numeric")
result <- sparse_mult(x, y)
}
