% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse Numeric Vector Class}
\description{
Sparse Numeric Vector Class
}
\section{Slots}{

\describe{
\item{\code{value}}{Numeric vector of non-zero values}

\item{\code{pos}}{Integer vector of positions of non-zero values}

\item{\code{length}}{Integer specifying the total length of the vector}
}}

