% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract two sparse numeric vectors}
\usage{
sparse_sub(x, y, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional arguments (not used)}
}
\value{
A sparse_numeric object containing the difference
}
\description{
Subtract two sparse numeric vectors
}
\section{Functions}{
\itemize{
\item \code{sparse_sub(x = sparse_numeric, y = sparse_numeric)}: Method for sparse_numeric vectors

}}
\examples{
x <- as(c(5, 0, 3, 0, 1), "sparse_numeric")
y <- as(c(1, 0, 2, 0, 1), "sparse_numeric")
result <- sparse_sub(x, y)
}
