% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{standardize,sparse_numeric-method}
\alias{standardize,sparse_numeric-method}
\title{Standardize sparse numeric vector}
\usage{
\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments (not used)}
}
\value{
A standardized sparse_numeric object
}
\description{
Standardize sparse numeric vector
}
