% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_functions.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector}
\usage{
standardize(x, ...)
}
\arguments{
\item{x}{An object to be standardized}

\item{...}{Additional arguments passed to methods}
}
\value{
A standardized object of the same class
}
\description{
Standardize a vector
}
