% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{*,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Multiply two sparse vectors together}
\usage{
\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{The first sparse vector}

\item{e2}{The second sparse vector, of the same length as the first}
}
\value{
Sparse vector of the product of the first vector x the second vector
}
\description{
This function combines two sparse vectors of the same length, and then returns a sparse vector of their product
}
