% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{cross product of two sparse vectors}
\usage{
sparse_crossprod(first, second, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(first, second)
}
\arguments{
\item{first}{The first sparse vector}

\item{second}{The second sparse vector, of the same length as the first}

\item{...}{Not Necessary}
}
\value{
numeric vector of a length of 1
}
\description{
This function calculates the cross product of two sparse vectors, and returns the numeric sum of the sparse vector's cross product
}
