% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{subtract one sparse vector from another}
\usage{
sparse_sub(first, second, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(first, second)
}
\arguments{
\item{first}{The first sparse vector}

\item{second}{The second sparse vector, of the same length as the first, that is subtracting from the first}

\item{...}{Not Necessary}
}
\value{
Sparse vector of the result after subtraction
}
\description{
This function subtracts one sparse vector, of the same length, by another sparse vector, and then returns the remainder left of the original sparse vector
}
